Added suggestion by Keith and fixed the issues
